feat: uniform style for all intros
