+++ src/screens/Game/TriviaIntro.jsx
@@ -1,52 +1,11 @@
 import React from "react";
-import { View, StyleSheet, Image } from "react-native";
-import { Button } from "react-native-elements";
+import { View, Image } from "react-native";
 import PropTypes from "prop-types";
 import Text from "../../components/Text";
+import Button from "../../components/Button";
+import styles from '../../styles/intro';
 
-const styles = StyleSheet.create({
-  root: {
-    flex: 1,
-    alignItems: "center",
-    justifyContent: "space-between",
-    backgroundColor: "white",
-    padding: 25,
-    paddingVertical: 60,
-  },
-  instructions: {
-    fontSize: 18,
-    textAlign: "center",
-    marginVertical: 20,
-  },
-  headInstruction: {
-    fontSize: 24,
-    fontWeight: "bold",
-    textAlign: "center",
-  },
-  nextButton: {
-    marginTop: 20,
-    width: 320,
-    height: 55,
-    borderRadius: 5,
-    backgroundColor: "#005AA3",
-  },
-  imageContainer: {
-    flex: 2,
-    width: "100%",
-    height: 200,
-    borderRadius: 2,
-    flexDirection: "column",
-    padding: 8,
-  },
-  image: {
-    ...StyleSheet.absoluteFillObject,
-    width: undefined,
-    height: undefined,
-    resizeMode: "contain",
-  },
-});
-
-const image = require("../../assets/writing_icon.png");
+const image = require("../../assets/books.png");
 
 function TriviaIntro({ navigation }) {
   return (
@@ -56,7 +15,7 @@ function TriviaIntro({ navigation }) {
       </View>
       <View style={styles.instructions}>
         <Text style={styles.headInstruction}>Grab a pencil and some paper.</Text>
-        <Text style={styles.instructions}>Writing is a great way to exercise your brain so please write each trivia question first before answering.</Text>
+        <Text style={styles.subInstructions}>Writing is a great way to exercise your brain so please write each trivia question first before answering.</Text>
         <Text style={styles.instructions}>Total time: 5 minutes</Text>
       </View>
       <Button
