+++ src/screens/Game/WritingIntro.jsx
@@ -1,49 +1,9 @@
 import React from "react";
-import { View, StyleSheet, Image } from "react-native";
-import { Button } from "react-native-elements";
+import { View, Image } from "react-native";
 import PropTypes from "prop-types";
 import Text from "../../components/Text";
-
-const styles = StyleSheet.create({
-  root: {
-    flex: 1,
-    alignItems: "center",
-    justifyContent: "space-between",
-    margin: 25,
-    marginVertical: 60,
-  },
-  instructions: {
-    fontSize: 18,
-    textAlign: "center",
-    marginVertical: 20,
-  },
-  headInstruction: {
-    fontSize: 24,
-    fontWeight: "bold",
-    textAlign: "center",
-  },
-  nextButton: {
-    marginTop: 20,
-    width: 320,
-    height: 55,
-    borderRadius: 5,
-    backgroundColor: "#005AA3",
-  },
-  imageContainer: {
-    flex: 2,
-    width: "100%",
-    height: 200,
-    borderRadius: 2,
-    flexDirection: "column",
-    padding: 8,
-  },
-  image: {
-    ...StyleSheet.absoluteFillObject,
-    width: undefined,
-    height: undefined,
-    resizeMode: "contain",
-  },
-});
+import Button from "../../components/Button";
+import styles from '../../styles/intro';
 
 const image = require("../../assets/writing_icon.png");
